+++ Server/Models/Commands.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Server.src
+﻿namespace Server.Models
 {
     internal class Commands
     {
