+++ Server/Models/Models.md
@@ -1,0 +1,2 @@
+﻿# Models
+Models contains any definitions for structures, including abstract classes.
