+++ Server/Resources/Controls/Styles.xaml
@@ -1,7 +1,11 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-
-    <Style x:Key="WindowsButtonStyle" TargetType="Button">
+    
+    <BitmapImage x:Key="CloseButtonImage" UriSource="/Resources/Images/Close.png" />
+    <BitmapImage x:Key="MinimizeButtonImage" UriSource="/Resources/Images/Minimize.png" />
+    <BitmapImage x:Key="ShowPasswordButtonImage" UriSource="/Resources/Images/ShowPassword.png" />
+    
+    <Style x:Key="MinimizeButtonStyle" TargetType="Button">
         <Setter Property="BorderThickness" Value="0"/>
         <Setter Property="Background" Value="Transparent"/>
         <Setter Property="Template">
@@ -9,7 +13,7 @@
                 <ControlTemplate TargetType="Button">
                     <Grid>
                         <Border x:Name="ButtonBorder" Background="Transparent" BorderThickness="0"/>
-                        <Image Source="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=Tag}" Stretch="Uniform"/>
+                        <Image Source="{StaticResource MinimizeButtonImage}" Stretch="Uniform"/>
                     </Grid>
                     <ControlTemplate.Triggers>
                         <Trigger Property="IsMouseOver" Value="True">
@@ -32,7 +36,7 @@
                 <ControlTemplate TargetType="Button">
                     <Grid>
                         <Border x:Name="ButtonBorder" Background="Transparent" BorderThickness="0"/>
-                        <Image Source="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=Tag}" Stretch="Uniform"/>
+                        <Image Source="{StaticResource CloseButtonImage}" Stretch="Uniform"/>
                     </Grid>
                     <ControlTemplate.Triggers>
                         <Trigger Property="IsMouseOver" Value="True">
@@ -46,7 +50,7 @@
             </Setter.Value>
         </Setter>
     </Style>
-
+    
     <Style TargetType="TextBox" x:Key="TextBoxStyle">
         <Setter Property="BorderThickness" Value="1"/>
         <Setter Property="Template">
