+++ Server/Resources/Resources.md
@@ -1,0 +1,6 @@
+﻿# Resource Director
+This directory contains static information, look up tables, static xaml definitions, or things that are
+not actively processes by SRC Server.
+
+The `Scripts` directory is here because it is only used during installation to a DCS Server and is not 
+part of the code executed by SRC Server.
