+++ Server/ViewModels/Utils/LoadFile.cs
@@ -1,12 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Server.Models;
+﻿using System.IO;
 
-namespace Server.Models
+namespace Server.ViewModels.Utils
 {
     public class LoadFile
     {
