+++ Server/ViewModels/ViewModel.md
@@ -1,0 +1,2 @@
+﻿# ViewModels
+All classes that are executed are placed here.
